requirements.txt
aiogram==3.4.1
yt-dlp
python-dotenv==1.0.0
